Se agregan las funcionalidades para compras, pero no se trae la información de los items de forma adecuada

+++ src/main/java/com/example/persistance/entity/Cliente.java
@@ -16,7 +16,7 @@ public class Cliente {
 
     private String apellidos;
 
-    private Integer celular;
+    private Long celular;
 
     public List<Compra> getCompras() {
         return compras;
@@ -58,11 +58,11 @@ public class Cliente {
         this.apellidos = apellidos;
     }
 
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
